+++ Backend/werqout/src/main/java/com/werqout/werqout/models/Coach.java
@@ -35,7 +35,7 @@ public class Coach extends User {
      * @param rating the average rating this user has received
      * @param numRatings the amount of ratings this user has recieved
      */
-    public Coach(long id, String userName, String email, String password, String groupName, String gymName,
+    public Coach(int id, String userName, String email, String password, String groupName, String gymName,
             double rating, int numRatings) {
         super(id, userName, email, password);
         this.groupName = groupName;
